Refactor highlighters to use decorated token stream
The previous implementation applied several text replacement filters
one after the other and they could clash and cause all sorts of
nastyness. It was also difficult to apply escaping correctly and I
hope that this refactoring will make it easier for me to make the
code robust instead of fragile.