Refactor documentation link handling
The Lua lexer previously always matched expressions like "io.input" as
one identifier. This was a quick and dirty hack to add support for
documentation links in the Lua highlighter. I've decided to keep the
hack because it can save quite a lot of code however this behavior is
now disabled by default and can be enabled with join_identifiers=true
(the Lua highlighter of course enables this behavior by default).